Add object classes
Added an abstract object class and conrete player and enemy class.
Included images for the player, one type of enemy and a black background
image. Implemented left and right movement of the player that can be
controlled via pressing the left and right arrow keyboard key. Defined
the basic interface of an enemy factory class that will be responsible
for creating new enemy instances in the game.