Add LevelGenerator
Implemented a LevelGenerator class that creates the initial enemies at the start of a new level at each iteration and stores the current level number. At the start of each level 5 rows of 11 equidistently places enemies are created. With increasing level number the starting position of these enemies is lowered by a small offset.