Add laser interaction
Implemented an interaction between Laser objects and other visible objects. If a laser that is shot by the player hits an enemy, both the laser an the enemy are killed. If a laser that is shot by an enemy hits the player, the laser is killed and the player looses a life. If the player has no lives left the game is lost. Collision detection is implemented via pygames own functions.