BaseObject refactor
Split the BaseObject class into 2 classes. The reason for this is that
BaseObject should represent all visible objects in the game but the
Blockade objects can not move, and thus needs no speed attribute.
Therefore, the new BaseObject class behaves the same as the old one
except it does not have a speed attribute anymore. The class
MovableObject inherits from BaseObject and adds the speed attribute.