Add scoreboard
Added a preliminary version of a scoreboard into
the game. The scoreboard displays the current
score, that so far only depends on the number of
killed enemies, and the number of lives the player
has left. Also added a new asset for the player
object and removed a duplicated method in the
MovableObject class.