Readme update
Checked milestones that have been achieved and
added new one. Also added a paragraph about the
newly implemented environment for reinforcement
learning and inserted a gif that displays the
computer playing the game by choosing random
actions. Implemented a test for the environment
that checks the shapes of the observations from
the environments `step()` method.